文章接口修改

+++ MP/src/main/java/com/luoyu/yorozuya/controller/ArticleController.java
@@ -42,5 +42,5 @@ public interface ArticleController {
      * @param id 文章编号
      * @return 文章VO
      */
-    ArticleVO getArticle(Integer id);
+    ArticleVO getArticle(Long id);
 }
